\name{runTime}
\alias{voidRef-class}
\alias{$,RC++Reference-method}
\alias{[[,ExternalArray,numeric-method}
\alias{[[,ExternalArrayWithLength,numeric-method}
\title{Run-time support for programmatically generated bindings to native code}
\description{
These are methods that are currently taken directly from 
a package that provides run-time support for 
programmatically generated bindings to native code.
They will be documented separately in that package
as it evolves.
}
\keyword{programming}
